feat: close detailedCard on bg click, refactor
