+++ RS_React/src/App.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import './App.css';
-import Wrapper from './components/Wrapper';
+import Main from './components/Main/Main';
 import ErrorBoundary from './components/ErrorBoundary';
 import DetailedCard from './components/DetailedCard/DetailedCard';
 
@@ -9,11 +9,17 @@ const router = createBrowserRouter([
   {
     path: '/',
     errorElement: <ErrorBoundary />,
-    element: <Wrapper />,
+    element: <Main />,
     children: [
       {
         path: ':id',
         element: <DetailedCard />,
+        // children: [
+        //   {
+        //     path: '*',
+        //     action:
+        //   }
+        // ]
       },
     ],
   },
