+++ RS_React/src/components/DetailedCard/DetailedCard.css
@@ -1,7 +1,9 @@
 .detailed-card {
   position: relative;
+  top: -20px;
+  right: 0;
   min-width: 30vw;
-  max-height: 100vh;
+  min-height: 100%;
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -33,4 +35,12 @@
   border-radius: 50%;
   line-height: 20px;
   cursor: pointer;
+}
+
+.close-background {
+  position: fixed;
+  transform: translate(-65%);
+  width: 100%;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.5);
 }
