+++ RS_React/src/components/Main/Main.css
@@ -1,0 +1,3 @@
+.main {
+  display: flex;
+}
