+++ RS_React/src/components/Wrapper/Wrapper.tsx
@@ -1,7 +1,7 @@
 import { ReactNode, useState } from 'react';
-import { SearchForm, TypeSearchResponse } from './SearchForm';
-import SearchResults from './SearchResults';
-import Pagination from './Pagination/Pagination';
+import { SearchForm, TypeSearchResponse } from '../SearchForm';
+import SearchResults from '../SearchResults';
+import Pagination from '../Pagination/Pagination';
 
 export default function Wrapper(): ReactNode {
   const [searchData, setSearchData] = useState<TypeSearchResponse | null>(null);
