+++ RS_React/src/components/styles/SearchResults.css
@@ -18,4 +18,5 @@
   border: 1px gray solid;
   border-radius: 10px;
   box-shadow: 10px 5px 5px gray;
+  cursor: pointer;
 }
