+++ RS_React/src/data/data.ts
@@ -1,2 +1,3 @@
-const baseURL = 'https://pokeapi.co/api/v2/pokemon/';
-export default baseURL;
+export const baseURL = 'https://pokeapi.co/api/v2/pokemon/';
+
+export const defaultItemsPerPage = '8';
